REFACTOR(cors): deduplicate CORS origin helper
